Rename OutlineProcess to OutlineUpdates
