+++ packages/outline/src/core/index.js
@@ -14,7 +14,7 @@ export type {
   TextMutation,
 } from './OutlineEditor';
 export type {ViewModel, ParsedViewModel} from './OutlineViewModel';
-export type {View} from './OutlineProcess';
+export type {View} from './OutlineUpdates';
 export type {
   NodeKey,
   OutlineNode,
